requirements.txt
absl-py==0.9.0
asn1crypto==0.24.0
astor==0.8.1
backports.weakref==1.0.post1
CherryPy==8.9.1
cryptography==2.1.4
enum34==1.1.6
funcsigs==1.0.2
futures==3.3.0
gast==0.3.2
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.6
ipaddress==1.0.17
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
Markdown==3.1.1
mercurial==5.2.2
mock==3.0.5
numpy==1.16.6
Pillow==6.2.2
protobuf==3.11.2
psutil==5.4.2
pycrypto==2.6.1
pygobject==3.26.1
pyOpenSSL==17.5.0
python-apt==1.6.5+ubuntu0.2
pyxdg==0.25
PyYAML==3.12
repoze.lru==0.7
Routes==2.4.1
scipy==1.1.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.15.2
termcolor==1.1.0
virtualenv==16.7.9
WebOb==1.7.3
Werkzeug==0.16.0
wrapt==1.11.2
